requirements.txt
# Data manipulation
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.11.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0

# Signal processing
pywavelets>=1.4.0

# Jupyter
jupyter>=1.0.0
jupyterlab>=4.0.0
ipywidgets>=8.0.0

# Configuration
pyyaml>=6.0.0
python-dotenv>=1.0.0

# Utilities
tqdm>=4.65.0
openpyxl>=3.1.0

# Statistical analysis
statsmodels>=0.14.0
pingouin>=0.5.3

# Report generation
jinja2>=3.1.0
